created src/redux folder and moved all actions and reducers there. Moved  utils folder to src/utils
