+++ src/components/app/app.js
@@ -3,8 +3,8 @@ import AppHeader from '../app-header/app-header';
 import Main from '../main/main';
 import Modal from '../modal/modal';
 import { useDispatch, useSelector } from 'react-redux';
-import { closeModalWithIngredient } from '../../services/actions/modal';
-import { modalContentTypes } from '../../services/utils/constants';
+import { closeModalWithIngredient } from '../../redux/actions/modal';
+import { modalContentTypes } from '../../utils/constants';
 import IngredientDetails from '../modal/ingredient-details/ingredient-details';
 import OrderDetails from '../modal/order-details/order-details';
