+++ src/components/burger-constructor/component-list/component-list.js
@@ -3,8 +3,8 @@ import BurgerComponent from '../burger-component/burger-component';
 import DummyComponent from '../dummy-component/dummy-component';
 import { useSelector, useDispatch } from "react-redux";
 import { useDrop } from "react-dnd";
-import { addConstructorComponent } from "../../../services/actions/burger-constructor";
-import { dndTypes } from "../../../services/utils/constants";
+import { addConstructorComponent } from "../../../redux/actions/burger-constructor";
+import { dndTypes } from "../../../utils/constants";
 
 function ComponentList() {
     const dispatch = useDispatch();
