+++ src/redux/actions/burger-ingredients.js
@@ -1,4 +1,4 @@
-import { getIngredients } from "../utils/api";
+import { getIngredients } from "../../services/api";
 
 export const GET_INGREDIENTS_REQUEST = 'GET_INGREDIENTS_REQUEST';
 export const GET_INGREDIENTS_SUCCESS = 'GET_INGREDIENTS_SUCCESS';
