+++ src/redux/actions/modal.js
