+++ src/services/api.js
@@ -1,4 +1,4 @@
-import { BASE_URL } from "./constants";
+import { BASE_URL } from "../utils/constants";
 
 const getResponse = (response) => {
     if (response.ok) {
