+++ src/utils/constants.js
@@ -1,8 +1,10 @@
 export const BASE_URL = 'https://norma.nomoreparties.space/api';
+
 export const modalContentTypes = {
     INGREDIENT_DETAILS: 'INGREDIENT_DETAILS',
     ORDER_DETAILS: 'ORDER_DETAILS'
 }
+
 export const dndTypes = {
     INGREDIENT: 'INGREDIENT',
     CONSTRUCTOR_ITEM: 'CONSTRUCTOR_ITEM'
